astral/renderer/renderer: revamp curve clipping error reporting
Instead of a simple counter, provide a callback that give more
localized detailed information.

Error's detected all seem to be numerial round off where the values
expected to match are nearly matching (so matching that when printed
the values are printed the same). The algorithm recovers by adding
these tiny line segments.