astral/text/freetype_face_generator: work arounds for when Freetype reports a zero size for a COLR0 glyph
This issue is seen in the font from https://github.com/mozilla/twemoji-colr/.
Indeed, this issue: https://github.com/mozilla/twemoji-colr/issues/50
which dates back to 2019 is still active (so over 3 and half years)
where some browsers do not render anything either for the glyphs.

This code does the work around that if the size is reported as 0 for a COLR0
glyph, then we derived the layout offsets and size entirely from the union
of the bounding boxes of the layers. Seems to work.