astral/renderer/gl3/render_engine_gl3: support WEBGL_clip_cull_distance
The clipping extension on WebGL2 as changed to that the extension string to
look for is WEBGL_clip_cull_distance, but the enable is for the OpenGL ES
extension GL_ANGLE_clip_cull_distance.

At the time of this commit, we do not have a browser implementing the
extension, but hopefully time wil be kind.