Revert "astral/util/c_array: add method to test if two c_array<>'s overlap in memory"
This reverts commit 0b778936e7829be6267dffc1ee810d311f93f040.

The method though well intentioned, is not actually portable as it makes
the assumption that the integer values alone are correct to test for
memory overlap. Admittedly, finding a system that does not behand that
was may take effort or even heroic effort. However, I do not actually
trust it.